backend/requirements.txt
fastapi[all]
sqlalchemy
sqlmodel
alembic
asyncpg
stripe
openai
python-multipart
passlib
pyjwt
uvicorn[standard]
httpx
pytest
pytest-asyncio
pytest-mock
websockets
python-dotenv
posthog
psycopg2
python-jose
pillow
aiosmtplib
boto3
aiohttp
bcrypt
tiktoken